apps/api/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
python-multipart>=0.0.9,<1.0
torch>=2.2,<2.6
torchvision>=0.17,<0.21
diffusers>=0.27,<0.28
transformers>=4.39,<5.0
accelerate>=0.30,<1.0
safetensors>=0.4,<1.0
pillow>=10.3,<12.0
peft>=0.11,<1.0
